(tms99x8) Create a separation between MSX and VDP functions
