--- libsrc/video/tms9918/stdio/generic_console/tms9918_variables.asm
@@ -1,8 +1,0 @@
-
-    SECTION bss_clib
-
-    PUBLIC __tms9918_pattern_name
-    PUBLIC __tms9918_pattern_generator
-
-__tms9918_pattern_name:        defw 0
-__tms9918_pattern_generator:   defw 0
