+++ libsrc/video/tms9918/video_tms9918.lst
@@ -8,7 +8,6 @@ video/tms9918/obj/${TARGET}/msx_set_char_form
 video/tms9918/obj/${TARGET}/msx_set_sprite_mode
 video/tms9918/obj/${TARGET}/gen_msxbios
 video/tms9918/obj/${TARGET}/msx_put_sprite_16
-video/tms9918/obj/${TARGET}/gen_get_trigger
 video/tms9918/obj/${TARGET}/dithpat
 video/tms9918/obj/${TARGET}/gen_color
 video/tms9918/obj/${TARGET}/gen_vram
@@ -36,10 +35,10 @@ video/tms9918/obj/${TARGET}/msx_vpoke
 video/tms9918/obj/${TARGET}/gen_ldirmv
 video/tms9918/obj/${TARGET}/gen_set_mode
 video/tms9918/obj/${TARGET}/gen_vwrite
-video/tms9918/obj/${TARGET}/vdp_variables
 video/tms9918/obj/${TARGET}/msx_vmerge
 video/tms9918/obj/${TARGET}/msx_map_m2_pixel
 video/tms9918/obj/${TARGET}/vdpport
+video/tms9918/obj/${TARGET}/vdp_variables
 video/tms9918/obj/${TARGET}/msx_set_sprite_16
 video/tms9918/obj/${TARGET}/__tms9918_status_register
 video/tms9918/stdio/**/obj/${TARGET}/*
